+++ crates/aeronet_websocket/examples/websocket_client.rs
@@ -17,9 +17,9 @@ fn main() -> AppExit {
         .add_plugins((DefaultPlugins, EguiPlugin, WebSocketClientPlugin))
         .init_resource::<GlobalUi>()
         .add_systems(Update, (global_ui, add_msgs_to_ui, session_ui))
-        .observe(on_connecting)
-        .observe(on_connected)
-        .observe(on_disconnected)
+        .add_observer(on_connecting)
+        .add_observer(on_connected)
+        .add_observer(on_disconnected)
         .run()
 }
 
@@ -111,7 +111,7 @@ fn global_ui(mut egui: EguiContexts, mut commands: Commands, mut ui_state: ResMu
             let name = format!("{}. {target}", ui_state.session_id);
             commands
                 .spawn((Name::new(name), SessionUi::default()))
-                .add(WebSocketClient::connect(config, target));
+                .queue(WebSocketClient::connect(config, target));
         }
 
         for msg in &ui_state.log {