+++ crates/aeronet_webtransport/examples/webtransport_client.rs
@@ -21,9 +21,9 @@ fn main() -> AppExit {
         .add_plugins((DefaultPlugins, EguiPlugin, WebTransportClientPlugin))
         .init_resource::<GlobalUi>()
         .add_systems(Update, (global_ui, add_msgs_to_ui, session_ui))
-        .observe(on_connecting)
-        .observe(on_connected)
-        .observe(on_disconnected)
+        .add_observer(on_connecting)
+        .add_observer(on_connected)
+        .add_observer(on_disconnected)
         .run()
 }
 
@@ -132,7 +132,7 @@ fn global_ui(mut egui: EguiContexts, mut commands: Commands, mut ui_state: ResMu
                 let name = format!("{}. {target}", ui_state.session_id);
                 commands
                     .spawn((Name::new(name), SessionUi::default()))
-                    .add(WebTransportClient::connect(config, target));
+                    .queue(WebTransportClient::connect(config, target));
             }
         })();