+++ examples/echo_client/src/main.rs
@@ -57,9 +57,9 @@ fn main() -> AppExit {
             ui, // ..draw the UI for the session
         ))
         // Set up some observers to run when the session state changes
-        .observe(on_connecting)
-        .observe(on_connected)
-        .observe(on_disconnected)
+        .add_observer(on_connecting)
+        .add_observer(on_connected)
+        .add_observer(on_disconnected)
         .run()
 }
 
@@ -130,7 +130,7 @@ fn setup(mut commands: Commands) {
     ));
     // Make an `EntityCommand` via `connect`, which will set up this
     // session, and push that command onto the entity.
-    entity.add(WebSocketClient::connect(config, target));
+    entity.queue(WebSocketClient::connect(config, target));
 }
 
 // Observe state change events using `Trigger`s.