+++ examples/move_box_client/src/main.rs
@@ -50,9 +50,9 @@ fn main() -> AppExit {
                 (draw_boxes, handle_inputs).run_if(in_state(GameState::Playing)),
             ),
         )
-        .observe(on_connecting)
-        .observe(on_connected)
-        .observe(on_disconnected)
+        .add_observer(on_connecting)
+        .add_observer(on_connected)
+        .add_observer(on_disconnected)
         .run()
 }
 
@@ -74,7 +74,7 @@ struct WebSocketUi {
 }
 
 fn setup_level(mut commands: Commands) {
-    commands.spawn(Camera2dBundle::default());
+    commands.spawn(Camera2d);
 }
 
 fn on_connecting(
@@ -219,7 +219,7 @@ fn web_transport_ui(
             let name = format!("{}. {target}", global_ui.session_id);
             commands
                 .spawn((Name::new(name), AeronetRepliconClient))
-                .add(WebTransportClient::connect(config, target));
+                .queue(WebTransportClient::connect(config, target));
         }
     });
 }
@@ -315,7 +315,7 @@ fn web_socket_ui(
             let name = format!("{}. {target}", global_ui.session_id);
             commands
                 .spawn((Name::new(name), AeronetRepliconClient))
-                .add(WebSocketClient::connect(config, target));
+                .queue(WebSocketClient::connect(config, target));
         }
     });
 }
@@ -357,6 +357,6 @@ fn handle_inputs(mut inputs: EventWriter<PlayerInput>, input: Res<ButtonInput<Ke
 
 fn draw_boxes(mut gizmos: Gizmos, players: Query<(&PlayerPosition, &PlayerColor)>) {
     for (PlayerPosition(pos), PlayerColor(color)) in &players {
-        gizmos.rect(pos.extend(0.0), Quat::IDENTITY, Vec2::ONE * 50.0, *color);
+        gizmos.rect_2d(*pos, Vec2::ONE * 50.0, *color);
     }
 }